(piqi,erlang,ocaml) Take last field out of several when reading Protobuf
In Protocol Buffers binary format, fields defined as "required" or "optional"
can be present in the encoded record more that one time. The new behavior is to
take the last entry as the field's value (rather than taking the first).

The reason for this change is to follow Protocol Buffers specification more
closely.