(erlang) Optimize Protobuf records parsing to O(N)
Optimize parsing of Protocol Buffers records from O(N^2) to O(N) on the number
of fields. This applies to the number of steps and to the number of memory
allocations.

Worst case senario, when record fields are not ordered by numeric code, becomes
loglinear O(N * log(N)) instead of quadratic. N * log(N) steps are required to
sort the fields using mergesort.