Add support for a new .piq-positional field property
Just pick up the new self-spec from the development version of piqi.

When parsing records represented in Piq format, it allows to explicitly control
whether to accept a field if it is specified without a label (arguments without
a label are called "positional").

The default behavior is to allow fields of primitive types to be positional but
require field labels for records and lists.