Optionally preserve unknown Protobuf fields
Can be enabled by specifying piqic-erlang --gen-preserve-unknown-fields flag.

When enabled, unknown (unrecognized) Protobuf fields are captured during
de-serialization in a special 'piqi_unknown_pb' field and automatically written
back when the record is serialized to Protobuf.

Such treatment of unknown fields is consistent with Protocol Buffers behavior.
For instance, its manual states that "... the unknown fields are not discarded,
and if the message is later serialized, the unknown fields are serialized along
with it – so if the message is passed on to new code, the new fields are still
available.

It will be likely enabled by default in the next major release.